refactor(checker): split lib interface heritage merging into its own module
`lib_resolution.rs` crossed the 2000-line file cap after the heritage
resolution fix. Move `merge_lib_interface_heritage` and its two heritage
type-arg helpers into a new `lib_resolution_heritage` sibling module
(mechanical extraction, no behavior change), bringing both files under the cap.
The relocated doc block now correctly documents `resolve_lib_type_by_name`.
