+++ crates/tsz-checker/src/types/queries/mod.rs
@@ -26,6 +26,7 @@ mod lib_name_text;
 pub(crate) mod lib_namespace_direct;
 pub(crate) mod lib_prime;
 pub(crate) mod lib_resolution;
+pub(crate) mod lib_resolution_heritage;
 pub(crate) mod lib_resolution_selected;
 pub(crate) mod lib_scoped_heritage;
 pub(crate) mod type_only;
